Refactor: abstract additional Items::Find controller logic into Validatable module
